LLM: Remap night brightness 70% to 100%

+++ src/main/java/rs117/hd/renderer/SkyRenderer.java
@@ -67,7 +67,7 @@ public class SkyRenderer {
 	 */
 	private static final class Sample {
 		private Environment environment;
-		private int minBrightness;
+		private float minBrightness;
 		private int frame;
 		private float[] zenithSrgb;
 		private float[] horizonSrgb;
